\name{fqtl-package}
\alias{fqtl-package}
\alias{fqtl}
\docType{package}
\title{
\packageTitle{fqtl}
}
\description{
\packageDescription{fqtl}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fqtl}
\packageIndices{fqtl}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{fqtl}
Maintainer: \packageMaintainer{fqtl}
}
\references{Park, Sarkar, Butani, Kellis (2017)}

\keyword{package}
\seealso{}
\examples{}
