\name{I/O}
\alias{read.plink}
\title{Read binary PLINK format}
\description{
}
\usage{
read.plink(bed.header)
}
\arguments{
  \item{bed.header}{header file for PLINK binary format}
}
\value{
  \code{read.plink()} returns a list of FAM, BIM, BED data.
}
\details{
}
\references{
}
\examples{
}
\author{Yongjin Park}

