% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nb.normalize}
\alias{nb.normalize}
\title{Normalize Negative Binomial data by row-wise geometric mean}
\usage{
nb.normalize(Y)
}
\arguments{
\item{Y}{count data matrix}
}
\value{
normalized count data across columns
}
\description{
Normalize Negative Binomial data by row-wise geometric mean
}
