% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read.plink}
\alias{read.plink}
\title{Read binary PLINK format}
\usage{
read.plink(bed.header)
}
\arguments{
\item{bed.header}{header for plink fileset}
}
\value{
a list of FAM, BIM, BED data.
}
\description{
Read binary PLINK format
}
\author{
Yongjin Park, \email{ypp@csail.mit.edu}, \email{yongjin.peter.park@gmail.com}
}
